This is implicit

+++ src/multiply.cpp
@@ -61,7 +61,7 @@ struct Performer
         const float multiply = data->gains.front();
         Type* type = static_cast<Type*>(data->buffer.data);
         for (int i = 0; i < iterations; ++i, ++type) {
-            *type = std::clamp<int64_t>(static_cast<int64_t>(static_cast<int64_t>(*type) * multiply),
+            *type = std::clamp<int64_t>(static_cast<int64_t>(*type) * multiply,
                                         std::numeric_limits<Type>::min(),
                                         std::numeric_limits<Type>::max());
         }
@@ -92,7 +92,7 @@ struct Performer<24, true, Type>
             data8[3] = flag;
 
             // multiply and mask
-            data32 = std::clamp<int64_t>(static_cast<int64_t>(static_cast<int64_t>(data32) * multiply),
+            data32 = std::clamp<int64_t>(static_cast<int64_t>(data32) * multiply,
                                          std::numeric_limits<Type>::min(),
                                          std::numeric_limits<Type>::max());
             data32 &= 0x7fffff;
